internal/olm/client{,.go}: refactor Client to its own package
